refactor godotassetdelivery to godot.assetdelivery
