+++ plugin/PlayAssetDelivery/src/test/java/com/google/play/godot/ExampleUnitTest.java
@@ -14,29 +14,20 @@
   	limitations under the License.
  */
 
-package com.google.play.godotassetdelivery;
-
-import android.content.Context;
-
-import androidx.test.platform.app.InstrumentationRegistry;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
+package com.google.play.godot;
 
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import static org.junit.Assert.*;
 
 /**
- * Instrumented test, which will execute on an Android device.
+ * Example local unit test, which will execute on the development machine (host).
  *
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
-@RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest {
+public class ExampleUnitTest {
     @Test
-    public void useAppContext() {
-        // Context of the app under test.
-        Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
-        assertEquals("com.google.play.godotassetdelivery.test", appContext.getPackageName());
+    public void addition_isCorrect() {
+        assertEquals(4, 2 + 2);
     }
 }
